+++ locust/core.py
@@ -58,6 +58,35 @@ def task(weight=1):
         return decorator_func
 
 
+def seq_task(order):
+    """
+    Used as a convenience decorator to be able to declare tasks for a TaskSequence
+    inline in the class. Example::
+
+        class NormalUser(TaskSequence):
+            @seq_task(1)
+            def login_first(self):
+                pass
+
+            @seq_task(2)
+            @task(25) # You can also set the weight in order to execute the task for `weight` times one after another.
+            def then_read_thread(self):
+                pass
+
+            @seq_task(3)
+            def then_logout(self):
+                pass
+    """
+
+    def decorator_func(func):
+        func.locust_task_order = order
+        if not hasattr(func, 'locust_task_weight'):
+            func.locust_task_weight = 1
+        return func
+
+    return decorator_func
+
+
 class NoClientWarningRaiser(object):
     """
     The purpose of this class is to emit a sensible error message for old test scripts that 
@@ -87,6 +116,9 @@ class Locust(object):
     
     max_wait = 1000
     """Maximum waiting time between the execution of locust tasks"""
+
+    wait_function = lambda self: random.randint(self.min_wait,self.max_wait) 
+    """Function used to calculate waiting time between the execution of locust tasks in milliseconds"""
     
     task_set = None
     """TaskSet class that defines the execution behaviour of this locust"""
@@ -155,13 +187,22 @@ class HttpLocust(Locust):
     Instance of HttpSession that is created upon instantiation of Locust. 
     The client support cookies, and therefore keeps the session between HTTP requests.
     """
+
+    trust_env = False
+    """
+    Look for proxy settings will slow down the default http client.
+    It's the default behavior of the requests library.
+    We don't need this feature most of the time, so disable it by default.
+    """
     
     def __init__(self):
         super(HttpLocust, self).__init__()
         if self.host is None:
             raise LocustError("You must specify the base host. Either in the host attribute in the Locust class, or on the command line using the --host option.")
-        
-        self.client = HttpSession(base_url=self.host)
+
+        session = HttpSession(base_url=self.host)
+        session.trust_env = self.trust_env
+        self.client = session
 
 
 class TaskSetMeta(type):
@@ -204,9 +245,9 @@ class TaskSet(object):
     Class defining a set of tasks that a Locust user will execute. 
     
     When a TaskSet starts running, it will pick a task from the *tasks* attribute, 
-    execute it, call it's wait function which will sleep a random number between
-    *min_wait* and *max_wait* milliseconds. It will then schedule another task for 
-    execution and so on.
+    execute it, and call its *wait_function* which will define a time to sleep for. 
+    This defaults to a uniformly distributed random number between *min_wait* and 
+    *max_wait* milliseconds. It will then schedule another task for execution and so on.
     
     TaskSets can be nested, which means that a TaskSet's *tasks* attribute can contain 
     another TaskSet. If the nested TaskSet it scheduled to be executed, it will be 
@@ -246,6 +287,13 @@ class TaskSet(object):
     TaskSet.
     """
     
+    wait_function = None
+    """
+    Function used to calculate waiting time betwen the execution of locust tasks in milliseconds. 
+    Can be used to override the wait_function defined in the root Locust class, which will be used
+    if not set on the TaskSet.
+    """
+
     locust = None
     """Will refer to the root Locust class instance when the TaskSet has been instantiated"""
 
@@ -272,11 +320,13 @@ class TaskSet(object):
 
         self.parent = parent
         
-        # if this class doesn't have a min_wait or max_wait defined, copy it from Locust
+        # if this class doesn't have a min_wait, max_wait or wait_function defined, copy it from Locust
         if not self.min_wait:
             self.min_wait = self.locust.min_wait
         if not self.max_wait:
             self.max_wait = self.locust.max_wait
+        if not self.wait_function:
+            self.wait_function = self.locust.wait_function
 
         self._lock.acquire()
         if hasattr(self, "setup") and self._setup_has_run is False:
@@ -377,10 +427,12 @@ class TaskSet(object):
     def get_next_task(self):
         return random.choice(self.tasks)
     
+    def get_wait_secs(self):
+        millis = self.wait_function()
+        return millis / 1000.0
+
     def wait(self):
-        millis = random.randint(self.min_wait, self.max_wait)
-        seconds = millis / 1000.0
-        self._sleep(seconds)
+        self._sleep(self.get_wait_secs())
 
     def _sleep(self, seconds):
         gevent.sleep(seconds)
@@ -405,3 +457,34 @@ class TaskSet(object):
         Locust instance.
         """
         return self.locust.client
+
+
+class TaskSequence(TaskSet):
+    """
+    Class defining a sequence of tasks that a Locust user will execute.
+
+    When a TaskSequence starts running, it will pick the task in `index` from the *tasks* attribute,
+    execute it, and call its *wait_function* which will define a time to sleep for.
+    This defaults to a uniformly distributed random number between *min_wait* and
+    *max_wait* milliseconds. It will then schedule the `index + 1 % len(tasks)` task for execution and so on.
+
+    TaskSequence can be nested with TaskSet, which means that a TaskSequence's *tasks* attribute can contain
+    TaskSet instances as well as other TaskSequence instances. If the nested TaskSet is scheduled to be executed, it will be
+    instantiated and called from the current executing TaskSet. Execution in the
+    currently running TaskSet will then be handed over to the nested TaskSet which will
+    continue to run until it throws an InterruptTaskSet exception, which is done when
+    :py:meth:`TaskSet.interrupt() <locust.core.TaskSet.interrupt>` is called. (execution
+    will then continue in the first TaskSet).
+
+    In this class, tasks should be defined as a list, or simply define the tasks with the @seq_task decorator
+    """
+
+    def __init__(self, parent):
+        super(TaskSequence, self).__init__(parent)
+        self._index = 0
+        self.tasks.sort(key=lambda t: t.locust_task_order if hasattr(t, 'locust_task_order') else 1)
+
+    def get_next_task(self):
+        task = self.tasks[self._index]
+        self._index = (self._index + 1) % len(self.tasks)
+        return task