+++ locust/main.py
@@ -1,11 +1,12 @@
 import inspect
 import logging
 import os
+import importlib
 import signal
 import socket
 import sys
 import time
-from optparse import OptionParser
+import argparse
 
 import gevent
 
@@ -18,255 +19,238 @@ from .log import console_logger, setup_logging
 from .runners import LocalLocustRunner, MasterLocustRunner, SlaveLocustRunner
 from .stats import (print_error_report, print_percentile_stats, print_stats,
                     stats_printer, stats_writer, write_stat_csvs)
-from .util.time import parse_timespan
+from .util.timespan import parse_timespan
 
 _internals = [Locust, HttpLocust]
 version = locust.__version__
 
 def parse_options():
     """
-    Handle command-line options with optparse.OptionParser.
+    Handle command-line options with argparse.ArgumentParser.
 
     Return list of arguments, largely for use in `parse_arguments`.
     """
 
     # Initialize
-    parser = OptionParser(usage="locust [options] [LocustClass [LocustClass2 ... ]]")
+    parser = argparse.ArgumentParser()
 
-    parser.add_option(
+    parser.add_argument(
         '-H', '--host',
-        dest="host",
-        default=None,
         help="Host to load test in the following format: http://10.21.32.33"
     )
 
-    parser.add_option(
+    parser.add_argument(
         '--web-host',
-        dest="web_host",
         default="",
         help="Host to bind the web interface to. Defaults to '' (all interfaces)"
     )
     
-    parser.add_option(
+    parser.add_argument(
         '-P', '--port', '--web-port',
-        type="int",
-        dest="port",
+        type=int,
         default=8089,
         help="Port on which to run web host"
     )
     
-    parser.add_option(
+    parser.add_argument(
         '-f', '--locustfile',
-        dest='locustfile',
         default='locustfile',
         help="Python module file to import, e.g. '../other.py'. Default: locustfile"
     )
 
     # A file that contains the current request stats.
-    parser.add_option(
+    parser.add_argument(
         '--csv', '--csv-base-name',
-        action='store',
-        type='str',
         dest='csvfilebase',
-        default=None,
         help="Store current request stats to files in CSV format.",
     )
 
     # if locust should be run in distributed mode as master
-    parser.add_option(
+    parser.add_argument(
         '--master',
         action='store_true',
-        dest='master',
-        default=False,
         help="Set locust to run in distributed mode with this process as master"
     )
 
     # if locust should be run in distributed mode as slave
-    parser.add_option(
+    parser.add_argument(
         '--slave',
         action='store_true',
-        dest='slave',
-        default=False,
         help="Set locust to run in distributed mode with this process as slave"
     )
     
     # master host options
-    parser.add_option(
+    parser.add_argument(
         '--master-host',
-        action='store',
-        type='str',
-        dest='master_host',
         default="127.0.0.1",
         help="Host or IP address of locust master for distributed load testing. Only used when running with --slave. Defaults to 127.0.0.1."
     )
     
-    parser.add_option(
+    parser.add_argument(
         '--master-port',
-        action='store',
-        type='int',
-        dest='master_port',
+        type=int,
         default=5557,
         help="The port to connect to that is used by the locust master for distributed load testing. Only used when running with --slave. Defaults to 5557. Note that slaves will also connect to the master node on this port + 1."
     )
 
-    parser.add_option(
+    parser.add_argument(
         '--master-bind-host',
-        action='store',
-        type='str',
-        dest='master_bind_host',
         default="*",
         help="Interfaces (hostname, ip) that locust master should bind to. Only used when running with --master. Defaults to * (all available interfaces)."
     )
     
-    parser.add_option(
+    parser.add_argument(
         '--master-bind-port',
-        action='store',
-        type='int',
-        dest='master_bind_port',
+        type=int,
         default=5557,
         help="Port that locust master should bind to. Only used when running with --master. Defaults to 5557. Note that Locust will also use this port + 1, so by default the master node will bind to 5557 and 5558."
     )
 
-    parser.add_option(
+    parser.add_argument(
+        '--heartbeat-liveness',
+        type=int,
+        default=3,
+        help="set number of seconds before failed heartbeat from slave"
+    )
+
+    parser.add_argument(
+        '--heartbeat-interval',
+        type=int,
+        default=1,
+        help="set number of seconds delay between slave heartbeats to master"
+    )
+
+    parser.add_argument(
         '--expect-slaves',
-        action='store',
-        type='int',
-        dest='expect_slaves',
+        type=int,
         default=1,
         help="How many slaves master should expect to connect before starting the test (only when --no-web used)."
     )
 
     # if we should print stats in the console
-    parser.add_option(
+    parser.add_argument(
         '--no-web',
         action='store_true',
-        dest='no_web',
-        default=False,
         help="Disable the web interface, and instead start running the test immediately. Requires -c and -r to be specified."
     )
 
     # Number of clients
-    parser.add_option(
+    parser.add_argument(
         '-c', '--clients',
-        action='store',
-        type='int',
+        type=int,
         dest='num_clients',
         default=1,
         help="Number of concurrent Locust users. Only used together with --no-web"
     )
 
     # Client hatch rate
-    parser.add_option(
+    parser.add_argument(
         '-r', '--hatch-rate',
-        action='store',
-        type='float',
-        dest='hatch_rate',
+        type=float,
         default=1,
         help="The rate per second in which clients are spawned. Only used together with --no-web"
     )
     
     # Time limit of the test run
-    parser.add_option(
+    parser.add_argument(
         '-t', '--run-time',
-        action='store',
-        type='str',
-        dest='run_time',
-        default=None,
         help="Stop after the specified amount of time, e.g. (300s, 20m, 3h, 1h30m, etc.). Only used together with --no-web"
     )
-    
+
+    # skip logging setup
+    parser.add_argument(
+        '--skip-log-setup',
+        action='store_true',
+        dest='skip_log_setup',
+        default=False,
+        help="Disable Locust's logging setup. Instead, the configuration is provided by the Locust test or Python defaults."
+    )
+
     # log level
-    parser.add_option(
+    parser.add_argument(
         '--loglevel', '-L',
-        action='store',
-        type='str',
-        dest='loglevel',
         default='INFO',
         help="Choose between DEBUG/INFO/WARNING/ERROR/CRITICAL. Default is INFO.",
     )
     
     # log file
-    parser.add_option(
+    parser.add_argument(
         '--logfile',
-        action='store',
-        type='str',
-        dest='logfile',
-        default=None,
         help="Path to log file. If not set, log will go to stdout/stderr",
     )
     
     # if we should print stats in the console
-    parser.add_option(
+    parser.add_argument(
         '--print-stats',
         action='store_true',
-        dest='print_stats',
-        default=False,
         help="Print stats in the console"
     )
 
     # only print summary stats
-    parser.add_option(
+    parser.add_argument(
        '--only-summary',
        action='store_true',
-       dest='only_summary',
-       default=False,
        help='Only print the summary stats'
     )
 
-    parser.add_option(
+    parser.add_argument(
         '--no-reset-stats',
         action='store_true',
         help="[DEPRECATED] Do not reset statistics once hatching has been completed. This is now the default behavior. See --reset-stats to disable",
     )
 
-    parser.add_option(
+    parser.add_argument(
         '--reset-stats',
         action='store_true',
-        dest='reset_stats',
-        default=False,
         help="Reset statistics once hatching has been completed. Should be set on both master and slaves when running in distributed mode",
     )
     
     # List locust commands found in loaded locust files/source files
-    parser.add_option(
+    parser.add_argument(
         '-l', '--list',
         action='store_true',
         dest='list_commands',
-        default=False,
         help="Show list of possible locust classes and exit"
     )
     
     # Display ratio table of all tasks
-    parser.add_option(
+    parser.add_argument(
         '--show-task-ratio',
         action='store_true',
-        dest='show_task_ratio',
-        default=False,
         help="print table of the locust classes' task execution ratio"
     )
     # Display ratio table of all tasks in JSON format
-    parser.add_option(
+    parser.add_argument(
         '--show-task-ratio-json',
         action='store_true',
-        dest='show_task_ratio_json',
-        default=False,
         help="print json data of the locust classes' task execution ratio"
     )
     
     # Version number (optparse gives you --version but we have to do it
     # ourselves to get -V too. sigh)
-    parser.add_option(
+    parser.add_argument(
         '-V', '--version',
-        action='store_true',
-        dest='show_version',
-        default=False,
-        help="show program's version number and exit"
+        action='version',
+        version='%(prog)s {}'.format(version),
+    )
+
+    # set the exit code to post on errors
+    parser.add_argument(
+        '--exit-code-on-error',
+        type=int,
+        default=1,
+        help="sets the exit code to post on error"
+    )
+
+    parser.add_argument(
+        'locust_classes',
+        nargs='*',
+        metavar='LocustClass',
     )
 
     # Finalize
-    # Return three-tuple of parser + the output from parse_args (opt obj, args)
-    opts, args = parser.parse_args()
-    return parser, opts, args
+    # Return two-tuple of parser + the output from parse_args
+    return parser, parser.parse_args()
 
 
 def _is_package(path):
@@ -335,6 +319,22 @@ def load_locustfile(path):
     dictionary of ``{'name': callable}`` containing all callables which pass
     the "is a Locust" test.
     """
+
+    def __import_locustfile__(filename, path):
+        """
+        Loads the locust file as a module, similar to performing `import`
+        """
+        try:
+            # Python 3 compatible
+            source = importlib.machinery.SourceFileLoader(os.path.splitext(locustfile)[0], path)
+            imported = source.load_module()
+        except AttributeError:
+            # Python 2.7 compatible
+            import imp
+            imported = imp.load_source(os.path.splitext(locustfile)[0], path)
+
+        return imported
+
     # Get directory and locustfile name
     directory, locustfile = os.path.split(path)
     # If the directory isn't in the PYTHONPATH, add it so our import will work
@@ -354,8 +354,8 @@ def load_locustfile(path):
             # Add to front, then remove from original position
             sys.path.insert(0, directory)
             del sys.path[i + 1]
-    # Perform the import (trimming off the .py)
-    imported = __import__(os.path.splitext(locustfile)[0])
+    # Perform the import
+    imported = __import_locustfile__(locustfile, path)
     # Remove directory from path if we added it ourselves (just to be neat)
     if added_to_path:
         del sys.path[0]
@@ -368,15 +368,13 @@ def load_locustfile(path):
     return imported.__doc__, locusts
 
 def main():
-    parser, options, arguments = parse_options()
+    parser, options = parse_options()
 
     # setup logging
-    setup_logging(options.loglevel, options.logfile)
+    if not options.skip_log_setup:
+        setup_logging(options.loglevel, options.logfile)
+
     logger = logging.getLogger(__name__)
-    
-    if options.show_version:
-        print("Locust %s" % (version,))
-        sys.exit(0)
 
     locustfile = find_locustfile(options.locustfile)
 
@@ -401,13 +399,13 @@ def main():
         sys.exit(1)
 
     # make sure specified Locust exists
-    if arguments:
-        missing = set(arguments) - set(locusts.keys())
+    if options.locust_classes:
+        missing = set(options.locust_classes) - set(locusts.keys())
         if missing:
             logger.error("Unknown Locust(s): %s\n" % (", ".join(missing)))
             sys.exit(1)
         else:
-            names = set(arguments) & set(locusts.keys())
+            names = set(options.locust_classes) & set(locusts.keys())
             locust_classes = [locusts[n] for n in names]
     else:
         # list() call is needed to consume the dict_view object in Python 3
@@ -519,10 +517,7 @@ def main():
         main_greenlet.join()
         code = 0
         if len(runners.locust_runner.errors) or main_greenlet.exception:
-            code = 1
+            code = options.exit_code_on_error
         shutdown(code=code)
     except KeyboardInterrupt as e:
         shutdown(0)
-
-if __name__ == '__main__':
-    main()