+++ locust/stats.py
@@ -103,7 +103,7 @@ class RequestStats(object):
         if not entry:
             entry = StatsError(method, name, error)
             self.errors[key] = entry
-        entry.occured()
+        entry.occurred()
     
     def get(self, name, method):
         """
@@ -121,6 +121,7 @@ class RequestStats(object):
         """
         self.start_time = time.time()
         self.total.reset()
+        self.errors = {}
         for r in six.itervalues(self.entries):
             r.reset()
     
@@ -276,7 +277,7 @@ class StatsEntry(object):
     @property
     def fail_ratio(self):
         try:
-            return float(self.num_failures) / (self.num_requests + self.num_failures)
+            return float(self.num_failures) / self.num_requests
         except ZeroDivisionError:
             if self.num_failures > 0:
                 return 1.0
@@ -294,8 +295,18 @@ class StatsEntry(object):
     def median_response_time(self):
         if not self.response_times:
             return 0
+        median = median_from_dict(self.num_requests, self.response_times)
 
-        return median_from_dict(self.num_requests, self.response_times)
+        # Since we only use two digits of precision when calculating the median response time 
+        # while still using the exact values for min and max response times, the following checks 
+        # makes sure that we don't report a median > max or median < min when a StatsEntry only 
+        # have one (or very few) really slow requests
+        if median > self.max_response_time:
+            median = self.max_response_time
+        elif median < self.min_response_time:
+            median = self.min_response_time
+
+        return median
 
     @property
     def current_rps(self):
@@ -332,7 +343,11 @@ class StatsEntry(object):
         self.num_failures = self.num_failures + other.num_failures
         self.total_response_time = self.total_response_time + other.total_response_time
         self.max_response_time = max(self.max_response_time, other.max_response_time)
-        self.min_response_time = min(self.min_response_time or 0, other.min_response_time or 0) or other.min_response_time
+        if self.min_response_time is not None and other.min_response_time is not None:
+            self.min_response_time = min(self.min_response_time, other.min_response_time)
+        elif other.min_response_time is not None:
+            # this means self.min_response_time is None, so we can safely replace it
+            self.min_response_time = other.min_response_time
         self.total_content_length = self.total_content_length + other.total_content_length
 
         for key in other.response_times:
@@ -383,11 +398,8 @@ class StatsEntry(object):
         return report
 
     def __str__(self):
-        try:
-            fail_percent = (self.num_failures/float(self.num_requests + self.num_failures))*100
-        except ZeroDivisionError:
-            fail_percent = 0
-        
+        fail_percent = self.fail_ratio * 100
+
         return (" %-" + str(STATS_NAME_WIDTH) + "s %7d %12s %7d %7d %7d  | %7d %7.2f") % (
             (self.method and self.method + " " or "") + self.name,
             self.num_requests,
@@ -484,11 +496,11 @@ class StatsEntry(object):
 
 
 class StatsError(object):
-    def __init__(self, method, name, error, occurences=0):
+    def __init__(self, method, name, error, occurrences=0):
         self.method = method
         self.name = name
         self.error = error
-        self.occurences = occurences
+        self.occurrences = occurrences
 
     @classmethod
     def parse_error(cls, error):
@@ -509,8 +521,8 @@ class StatsError(object):
         key = "%s.%s.%r" % (method, name, StatsError.parse_error(error))
         return hashlib.md5(key.encode('utf-8')).hexdigest()
 
-    def occured(self):
-        self.occurences += 1
+    def occurred(self):
+        self.occurrences += 1
 
     def to_name(self):
         return "%s %s: %r" % (self.method, 
@@ -521,7 +533,7 @@ class StatsError(object):
             "method": self.method,
             "name": self.name,
             "error": StatsError.parse_error(self.error),
-            "occurences": self.occurences
+            "occurrences": self.occurrences
         }
 
     @classmethod
@@ -530,7 +542,7 @@ class StatsError(object):
             data["method"], 
             data["name"], 
             data["error"], 
-            data["occurences"]
+            data["occurrences"]
         )
 
 
@@ -554,10 +566,11 @@ global_stats = RequestStats()
 A global instance for holding the statistics. Should be removed eventually.
 """
 
-def on_request_success(request_type, name, response_time, response_length):
+def on_request_success(request_type, name, response_time, response_length, **kwargs):
     global_stats.log_request(request_type, name, response_time, response_length)
 
-def on_request_failure(request_type, name, response_time, exception):
+def on_request_failure(request_type, name, response_time, exception, **kwargs):
+    global_stats.log_request(request_type, name, response_time, 0)
     global_stats.log_error(request_type, name, exception)
 
 def on_report_to_master(client_id, data):
@@ -578,7 +591,7 @@ def on_slave_report(client_id, data):
         if error_key not in global_stats.errors:
             global_stats.errors[error_key] = StatsError.from_dict(error)
         else:
-            global_stats.errors[error_key].occurences += error["occurences"]
+            global_stats.errors[error_key].occurrences += error["occurrences"]
     
     # save the old last_request_timestamp, to see if we should store a new copy
     # of the response times in the response times cache
@@ -645,7 +658,7 @@ def print_error_report():
     console_logger.info(" %-18s %-100s" % ("# occurrences", "Error"))
     console_logger.info("-" * (80 + STATS_NAME_WIDTH))
     for error in six.itervalues(global_stats.errors):
-        console_logger.info(" %-18i %-100s" % (error.occurences, error.to_name()))
+        console_logger.info(" %-18i %-100s" % (error.occurrences, error.to_name()))
     console_logger.info("-" * (80 + STATS_NAME_WIDTH))
     console_logger.info("")
 
@@ -732,4 +745,26 @@ def distribution_csv():
         else:
             rows.append('"%s",0,"N/A","N/A","N/A","N/A","N/A","N/A","N/A","N/A","N/A"' % s.name)
 
+    return "\n".join(rows)
+
+def failures_csv():
+    """"Return the contents of the 'failures' tab as a CSV."""
+    from . import runners
+
+    rows = [
+        ",".join((
+            '"Method"',
+            '"Name"',
+            '"Error"',
+            '"Occurrences"',
+        ))
+    ]
+
+    for s in sort_stats(runners.locust_runner.stats.errors):
+        rows.append('"%s","%s","%s",%i' % (
+            s.method,
+            s.name,
+            s.error,
+            s.occurrences,
+        ))
     return "\n".join(rows)